requirements.txt
streamlit==1.38.0
sentence-transformers==2.7.0
spacy==3.7.2
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
pdfplumber==0.11.2
python-docx==1.1.0
nltk==3.8.1
pytest==8.3.2
requests==2.31.0
scikit-learn==1.3.2
textblob==0.17.1
wordcloud==1.9.2
seaborn==0.12.2
matplotlib==3.8.2
openai==1.3.0
python-dotenv==1.0.0